+++ lib/browser/aliases.rb
@@ -15,10 +15,10 @@ module Browser
 
     DEVICE_ALIASES = %w[
       blackberry_playbook? console? ipad? iphone? ipod_touch? kindle?
-      kindle_fire? mobile? nintendo? nintendo_wii? nintendo_wiiu? playbook?
-      playstation3? playstation4? playstation? playstation_vita? ps3? ps4? psp?
-      psp_vita? silk? surface? tablet? tv? vita? wii? wiiu? xbox? xbox_360?
-      xbox_one?
+      kindle_fire? mobile? nintendo? nintendo_switch? nintendo_wii?
+      nintendo_wiiu? playbook? playstation3? playstation4? playstation?
+      playstation_vita? ps3? ps4? psp? psp_vita? silk? surface? tablet? tv?
+      vita? wii? wiiu? xbox? xbox_360? xbox_one?
     ].freeze
 
     def self.included(target)